Rename REST_Controller and add DI for data manager and validator
